+++ frontend/src/components/Collections/components/CollectionsGrid/components/CollectionRow/components/Popover/index.tsx
@@ -1,0 +1,45 @@
+import { Classes, Position } from "@blueprintjs/core";
+import React from "react";
+import { LeftAlignedDetailsCell } from "../common/style";
+import { FieldValues, StyledButton, StyledPopover } from "./style";
+
+const conditionalPopover = (values: string[]) => {
+  if (!values || values.length === 0) {
+    return <LeftAlignedDetailsCell>-</LeftAlignedDetailsCell>;
+  }
+
+  return (
+    <LeftAlignedDetailsCell>
+      <FieldValues>
+        {values[0]}
+        <br />
+        {values[1]}
+      </FieldValues>
+      {values.length > 2 && (
+        <StyledPopover
+          boundary="window"
+          modifiers={{
+            hide: { enabled: false },
+            preventOverflow: { enabled: false },
+          }}
+          position={Position.BOTTOM}
+          popoverClassName={Classes.POPOVER_CONTENT_SIZING}
+          content={
+            <div>
+              {values.map((val, idx) => (
+                <React.Fragment key={val}>
+                  {val}
+                  {idx !== values.length - 1 && <br />}
+                </React.Fragment>
+              ))}
+            </div>
+          }
+        >
+          <StyledButton minimal>+{values.length - 2}</StyledButton>
+        </StyledPopover>
+      )}
+    </LeftAlignedDetailsCell>
+  );
+};
+
+export default conditionalPopover;
