+++ frontend/src/components/Collections/components/CollectionsGrid/style.ts
@@ -1,11 +1,8 @@
 import { HTMLTable } from "@blueprintjs/core";
 import styled, { css } from "styled-components";
+import { detailsColWidthCSS } from "./components/common/style";
 
-export const detailsColWidthCSS = css`
-  width: calc(1 / 8 * 100%);
-`;
-
-export const titleColWidthCSS = css`
+const titleColWidthCSS = css`
   width: calc(3 / 8 * 100%);
 `;
 
@@ -14,10 +11,6 @@ export const StyledCollectionsGrid = styled(HTMLTable)`
   margin-top: 16px;
 `;
 
-export const CollectionDataCell = styled.td`
-  text-align: left !important;
-  ${detailsColWidthCSS}
-`;
 export const CollectionHeaderCell = styled.th`
   ${titleColWidthCSS}
   text-align: left !important;
@@ -27,11 +20,6 @@ export const LeftAlignedHeaderCell = styled.th`
   ${detailsColWidthCSS}
   text-align: left !important;
   margin-left: 16px;
-`;
-
-export const RightAlignedDataCell = styled.td`
-  width: calc(1 / 8 * 100%);
-  text-align: right !important;
 `;
 
 export const RightAlignedHeaderCell = styled.th`
